refactor(support): rename templates to support-templates and use react-icons
