+++ src/App.tsx
@@ -78,7 +78,7 @@ const SupportDashboardScreen = lazy(() => import('./screens/support-dashboard.sc
 const SupportDashboardIssueScreen = lazy(() => import('./screens/support-dashboard-issue.screen'));
 const SupportDashboardCreateScreen = lazy(() => import('./screens/support-dashboard-create.screen'));
 const NotesScreen = lazy(() => import('./screens/notes.screen'));
-const TemplatesScreen = lazy(() => import('./screens/templates.screen'));
+const TemplatesScreen = lazy(() => import('./screens/support-templates.screen'));
 const RealunitScreen = lazy(() => import('./screens/realunit.screen'));
 const RealunitHoldersScreen = lazy(() => import('./screens/realunit-holders.screen'));
 const RealunitQuotesScreen = lazy(() => import('./screens/realunit-quotes.screen'));
