+++ src/screens/support-dashboard-issue.screen.tsx
@@ -4,8 +4,8 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { FilePreviewPanel } from 'src/components/compliance/file-preview-panel';
 import { ErrorHint } from 'src/components/error-hint';
-import { TemplateArrayPickerModal } from 'src/components/templates/template-array-picker-modal';
-import { TemplatePickerModal } from 'src/components/templates/template-picker-modal';
+import { TemplateArrayPickerModal } from 'src/components/support-templates/template-array-picker-modal';
+import { TemplatePickerModal } from 'src/components/support-templates/template-picker-modal';
 import { useSettingsContext } from 'src/contexts/settings.context';
 import { TransactionInfo, UserDataDetail, useCompliance } from 'src/hooks/compliance.hook';
 import { useSupportDashboardGuard } from 'src/hooks/guard.hook';
